Shiny App
It has 4 tabs with the first one showing the location of the business on a map with clickable points that will show the name and average stars for the business. There is a table that shows the general information about the top business for each city. The fourth tab is contact information. The other tabs will need to filled in.